refactor: auth provider endpoint
