+++ src/main/resources/extensions/role-template-setting.yaml
@@ -14,6 +14,9 @@ rules:
   - apiGroups: [ "" ]
     resources: [ "settings" ]
     verbs: [ "create", "patch", "update", "delete", "deletecollection" ]
+  - apiGroups: [ "api.console.halo.run" ]
+    resources: [ "auth-providers/enable", "auth-providers/disable" ]
+    verbs: [ "update" ]
 ---
 apiVersion: v1alpha1
 kind: "Role"
@@ -29,4 +32,7 @@ metadata:
 rules:
   - apiGroups: [ "" ]
     resources: [ "settings" ]
+    verbs: [ "get", "list" ]
+  - apiGroups: [ "api.console.halo.run" ]
+    resources: [ "auth-providers" ]
     verbs: [ "get", "list" ]
